@@ -1,7 +1,8 @@
-namespace recipefinder.Models {
-    public class ErrorViewModel {
-        public string? RequestId { get; set; }
+namespace recipeFinder.Models;
 
-        public bool ShowRequestId => !string.IsNullOrEmpty(RequestId);
-    }
+public class ErrorViewModel
+{
+    public string? RequestId { get; set; }
+
+    public bool ShowRequestId => !string.IsNullOrEmpty(RequestId);
 }
