@@ -3,16 +3,8 @@
 }
 
 <div class="text-center">
-    <main class="row flex-row-reverse m-auto shadow bg-white px-0 rounded-3">
-        <div class="col">
-            <h2 class="text-center mb-5">Upload or take a photo</h2>
-            <!-- Todo: Create form username and password inputs -->
-            <form id="form">
-                <input type="file" accept="image/*" id="file-input">
-                <button type="submit" id="upload-button">Upload Image</button>
-            </form>
-            <img src="" id="preview" style="max-width: 800px; max-height: 500px;">
-        </div>
-    </main>
+    <h1 class="display-4">Welcome to Recipe Finder</h1>
+    <p>Upload a photo of the dish and we will send you the recipe</p>
+
+    <img src="https://miro.medium.com/max/1100/1*-ganvHfXEbn6oYk-krRpIg.webp">
 </div>
-<script src="validation.js"></script>
