@@ -1,3 +1,3 @@
-﻿@using recipefinder
-@using recipefinder.Models
+﻿@using recipeFinder
+@using recipeFinder.Models
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
